dask/requirements.txt
torchvision
dask
distributed
pillow
numpy